app_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
# Add testing libraries later: pytest, httpx
pytest
httpx
supabase-py
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
pypdf2
python-docx
openai
qdrant-client
fastapi-mail
